# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the administration API for the ImLate tracking system.",
        "title": "ImLate Admin API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
        "/admin-api/current-user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the user resolved from the JWT identity (`id`) in the request context.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "Get current authenticated admin user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_buzyka_imlate_internal_domain_entity.User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin-api/dashboard": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a simple dashboard welcome message for authenticated admin users.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-dashboard"
                ],
                "summary": "Get admin dashboard greeting",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_buzyka_imlate_internal_http_controller_adminapi.DashboardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_buzyka_imlate_internal_http_controller_adminapi.AuthErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_buzyka_imlate_internal_http_controller_adminapi.AuthErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin-api/reports/visits": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns paginated visitor attendance information aggregated per visitor per day for a given date range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-reports"
                ],
                "summary": "Get visitor attendance report",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD, inclusive)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD, inclusive)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by student flag",
                        "name": "is_student",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by year group",
                        "name": "year_group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by sign status (not_signed, signed_in, signed_out)",
                        "name": "sign_status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Records per page (default 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ReportsVisitsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a paginated, field-selectable visitor attendance report with multi-value filters and server-side sorting.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-reports"
                ],
                "summary": "Search visitor attendance report",
                "parameters": [
                    {
                        "description": "Search criteria",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.PostReportsVisitsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.PostReportsVisitsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin-api/theme": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns current tracking page theme assets and timing settings.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-theme"
                ],
                "summary": "Get current theme",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_buzyka_imlate_internal_usecase_theme.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin-api/theme/assets/{slot}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Uploads a custom asset for the given theme slot.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-theme"
                ],
                "summary": "Upload theme asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Theme slot",
                        "name": "slot",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Theme asset file",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_buzyka_imlate_internal_usecase_theme.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Removes a custom asset assignment and restores the default slot asset.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-theme"
                ],
                "summary": "Reset theme asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Theme slot",
                        "name": "slot",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_buzyka_imlate_internal_usecase_theme.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin-api/theme/settings": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates welcome and goodbye animation durations for the tracking page.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-theme"
                ],
                "summary": "Update theme settings",
                "parameters": [
                    {
                        "description": "Theme settings request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.UpdateThemeSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_buzyka_imlate_internal_usecase_theme.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin-api/track/visit": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Manually records a visitor sign-in/sign-out event on behalf of a visitor by an admin.\nUsed when a visitor forgot their key, lost it, or is otherwise unable to use it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-tracking"
                ],
                "summary": "Manual visitor track",
                "parameters": [
                    {
                        "description": "Track request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ManualTrackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ManualTrackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin-api/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all users available in the administration area.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "List admin users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_buzyka_imlate_internal_domain_entity.User"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new user account for the admin panel.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "Create admin user",
                "parameters": [
                    {
                        "description": "Create user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_buzyka_imlate_internal_domain_entity.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin-api/users/{id}/password": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Replaces the password of the user identified by UUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "Update user password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update password request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.UpdatePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin-api/visitors/{id}/key/{key}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Removes one access key from the specified visitor.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-visitors"
                ],
                "summary": "Remove visitor key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Visitor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Visitor key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_adminapi.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_buzyka_imlate_internal_domain_entity.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_buzyka_imlate_internal_domain_entity.UserRole"
                },
                "surname": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "internal_http_controller_adminapi.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT token in format: \"Bearer \u003cJWT\u003e\".\nToken lookup order: Authorization header, query parameter `token`, cookie `jwt`.\nProtected admin routes require admin role.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}